Security: Fix unlimited chat message size in SpacetimeDB
